requirements.txt
playwright
playwright-stealth
flask
gunicorn
python-dotenv
